fix(root): type fix
Typing fix in CodePreview component
